feat: Add AuditModel for created_at && updated_at && id field

+++ spring-data-tutorial/src/main/kotlin/com/example/springtutorial/springdatatutorial/repository/PostRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PostRepository: JpaRepository<Post, Int> {
+interface PostRepository: JpaRepository<Post, Long> {
 //    fun findAllByAuthorId(authorId: Int): List<Post> {
 //        return findAll().filter { it.author.id == authorId }
 //    }
